functions/load_vdb/requirements.txt
weaviate-client
pyyaml